collabwrk-backend/requirements.txt
fastapi[all]>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
openai>=1.10.0
python-socketio>=5.11.0
aiohttp>=3.9.1
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
